+++ pkg/dt/context.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package ctx
+package dt
 
 import (
 	"bytes"
